Move CloudSubscriber to separate class file.
